// HRMS-MATERIAL-UI/src/app/modules/AssociateExit/Models/causeanalysis.model.ts
export class CauseForm{
  
    employeeId: number;
    submitType:  string;
    associateExitId:  number;
    rootCause: string;
    actionItem: string;
    responsibility: string;
    tagretDate: string;
    actualDate: string;
    remarks: string;
    systemInfo: string;
    createdBy: string;
    modifiedBy:string;
  
  }